requirements.txt
# GUI Framework
PyQt6==6.6.1
PyQt6-Qt6==6.6.1
PyQt6-sip==13.6.0

# Packaging and Protection
pyinstaller==6.12.0
pyarmor==9.0.8

# Utilities
colorama==0.4.6
psutil==5.9.8
requests==2.31.0
PyJWT==2.8.0
DrissionPage
python-dotenv

# Development Tools (Optional)
black==24.1.1
pylint==3.0.3
pytest==8.0.0 